+++ docs/changelog.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Place any unreleased changes here, that are subject to release in coming versions :).
 
 * feat: Add structured logs to FastAPI (Uvicorn).
+* feat: Add structured logs to Django and Flask (Gunicorn).
 
 ## 2026-02-23